--- src/main/java/chess/domain/board/ChessBoardWrapper.java
@@ -1,27 +1,0 @@
-package chess.domain.board;
-
-import chess.domain.Position;
-import chess.domain.piece.Piece;
-import chess.domain.piece.PieceMoveResult;
-import chess.domain.piece.Team;
-import java.util.List;
-
-public class ChessBoardWrapper {
-    private final ChessBoard chessBoard;
-
-    public ChessBoardWrapper(ChessBoard chessBoard) {
-        this.chessBoard = chessBoard;
-    }
-
-    public PieceMoveResult move(Position from, Position to) {
-        return chessBoard.move(from, to);
-    }
-
-    public List<Piece> getPiecesOnBoard() {
-        return chessBoard.getPiecesOnBoard();
-    }
-
-    public double calculatePoint(Team team) {
-        return chessBoard.calculatePoint(team);
-    }
-}
