+++ src/main/java/chess/domain/piece/AbstractStraightMovePiece.java
@@ -2,7 +2,7 @@ package chess.domain.piece;
 
 
 import chess.domain.Position;
-import chess.domain.board.ChessBoard;
+import chess.domain.game.ChessBoard;
 import java.util.List;
 import java.util.Optional;
