+++ src/main/java/chess/domain/piece/King.java
@@ -1,7 +1,7 @@
 package chess.domain.piece;
 
 import chess.domain.Position;
-import chess.domain.board.ChessBoard;
+import chess.domain.game.ChessBoard;
 import java.util.Optional;
 
 public final class King extends AbstractCatchOnMovePiece {
