+++ src/test/java/chess/domain/piece/PawnTest.java
@@ -19,7 +19,7 @@ import static chess.domain.piece.Team.BLACK;
 import static chess.domain.piece.Team.WHITE;
 
 import chess.domain.Position;
-import chess.domain.board.ChessBoard;
+import chess.domain.game.ChessBoard;
 import java.util.List;
 import java.util.stream.Stream;
 import org.assertj.core.api.Assertions;
